requirements.txt
python-telegram-bot>=21.0
python-telegram-bot[socks]
fpdf2>=2.7.9
